Clean up code warnings, setting up basic vector math and mat44 struct

+++ engine/engine/Graphics/d3d12/D3d12Descriptor.cpp
@@ -38,7 +38,7 @@ void DescriptorHeap::Init()
 {
    D3D12_DESCRIPTOR_HEAP_DESC desc = {
       ToD3d12HeapType( mAllowedTypes ),
-      mHeapSize,
+      uint( mHeapSize ),
       mIsShaderVisible ? D3D12_DESCRIPTOR_HEAP_FLAG_SHADER_VISIBLE : D3D12_DESCRIPTOR_HEAP_FLAG_NONE
    };
